refactor!: extended Message interface
Similarly to the previous API, this allows users to send custom msgs. The difference is that destination, and source peers of the msg
are now part of the Message interface contract. This provides zero allocation way to send msgs! No internal msgWrapper allocations
for both Send and Broadcast methods.

Additionally, this is preparation for the new generics version.

Also includes minor test deflakes.